all_one/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
google-cloud-texttospeech==2.16.3
gtts==2.4.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6